+++ src/ui/SpaceWeatherPanel.h
@@ -31,7 +31,7 @@ public:
 
 private:
   void destroyCache();
-  static SDL_Color colorForK(int k, const ThemeColors &themes);
+  static SDL_Color colorForK(float k, const ThemeColors &themes);
   static SDL_Color colorForSFI(int sfi, const ThemeColors &themes);
   static SDL_Color colorForNOAAScale(int scale, const ThemeColors &themes);
